refactor: テストコードの重複排除と実装詳細のテスト削除
- 重複するテストフィクスチャ（sample_image_path, dark_image_pathなど）をconftest.pyに統合
- テスト画像作成ヘルパー関数_create_test_imageを共通化して6ファイルから削除
- 実装詳細をテストする以下のテストを削除
  - test_init_worker_initializes_class_variable: プライベートクラス変数の初期化テスト
  - test_get_image_features_batch_mode_returns_correct_shape: モック挙動に依存したテスト
- テストカバレッジを維持したまま95件から94件に削減
